tests: Combine Dynamic Rendering tests
