--- tests/positive/dynamic_rendering.cpp
@@ -1,1387 +1,0 @@
-/*
- * Copyright (c) 2015-2023 The Khronos Group Inc.
- * Copyright (c) 2015-2023 Valve Corporation
- * Copyright (c) 2015-2023 LunarG, Inc.
- * Copyright (c) 2015-2023 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-
-#include "../framework/layer_validation_tests.h"
-#include "generated/vk_extension_helper.h"
-
-#include <algorithm>
-#include <array>
-#include <chrono>
-
-class PositiveDynamicRendering : public VkPositiveLayerTest {
-  public:
-    void InitBasicDynamicRendering();
-};
-
-void PositiveDynamicRendering::InitBasicDynamicRendering() {
-    SetTargetApiVersion(VK_API_VERSION_1_1);
-    AddRequiredExtensions(VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (DeviceValidationVersion() < VK_API_VERSION_1_1) {
-        GTEST_SKIP() << "At least Vulkan version 1.1 is required";
-    }
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto dynamic_rendering_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingFeaturesKHR>();
-    GetPhysicalDeviceFeatures2(dynamic_rendering_features);
-    if (!dynamic_rendering_features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering , skipping.";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &dynamic_rendering_features, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT));
-}
-
-TEST_F(PositiveDynamicRendering, BasicUsage) {
-    TEST_DESCRIPTION("Most simple way to use dynamic rendering");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.layerCount = 1;
-    begin_rendering_info.renderArea = {{0, 0}, {64, 64}};
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, Draw) {
-    TEST_DESCRIPTION("Draw with Dynamic Rendering.");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_formats = VK_FORMAT_UNDEFINED;
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-
-    pipe.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    VkRect2D scissor = {{0, 0}, {16, 16}};
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, DrawMultiBind) {
-    TEST_DESCRIPTION("Draw with Dynamic Rendering and multiple CmdBindPipeline calls.");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    const auto depth_format = FindSupportedDepthOnlyFormat(gpu());
-    auto ds_ci = LvlInitStruct<VkPipelineDepthStencilStateCreateInfo>();
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.SetDepthStencil(&ds_ci);
-    pipe.AddDefaultColorAttachment();
-
-    VkPipelineObj pipe2(m_device);
-    pipe2.AddShader(&vs);
-    pipe2.AddShader(&fs);
-    pipe2.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_formats = VK_FORMAT_UNDEFINED;
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-    pipeline_rendering_info.depthAttachmentFormat = depth_format;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-    pipe.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    pipeline_rendering_info.depthAttachmentFormat = VK_FORMAT_UNDEFINED;
-    pipe2.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-    pipe2.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    VkRect2D scissor = {{0, 0}, {16, 16}};
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    // NOTE: Setting dynamic state does not count as "using" the currently bound pipeline.
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe2.handle());
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, BeginQuery) {
-    TEST_DESCRIPTION("Test calling vkCmdBeginQuery with a dynamic render pass.");
-
-    SetTargetApiVersion(VK_API_VERSION_1_3);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_3) {
-        GTEST_SKIP() << "At least Vulkan version 1.3 is required";
-    }
-
-    auto vk13features = LvlInitStruct<VkPhysicalDeviceVulkan13Features>();
-    auto features2 = GetPhysicalDeviceFeatures2(vk13features);
-    if (!vk13features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering";
-    }
-    if (!vk13features.synchronization2) {
-        GTEST_SKIP() << "Test requires (unsupported) synchronization2";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-    ASSERT_NO_FATAL_FAILURE(InitRenderTarget());
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.layerCount = 1;
-
-    VkQueryPoolCreateInfo qpci = LvlInitStruct<VkQueryPoolCreateInfo>();
-    qpci.queryType = VK_QUERY_TYPE_OCCLUSION;
-    qpci.queryCount = 2;
-
-    vk_testing::QueryPool query_pool;
-    query_pool.init(*m_device, qpci);
-
-    m_commandBuffer->begin();
-
-    vk::CmdBeginRendering(m_commandBuffer->handle(), &begin_rendering_info);
-    vk::CmdBeginQuery(m_commandBuffer->handle(), query_pool.handle(), 0, 0);
-    vk::CmdEndQuery(m_commandBuffer->handle(), query_pool.handle(), 0);
-    vk::CmdEndRendering(m_commandBuffer->handle());
-
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, PipeWithDiscard) {
-    TEST_DESCRIPTION("Create dynamic rendering pipeline with rasterizer discard.");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragShaderText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineDepthStencilStateCreateInfo ds_ci = LvlInitStruct<VkPipelineDepthStencilStateCreateInfo>();
-    ds_ci.depthTestEnable = VK_TRUE;
-    ds_ci.depthWriteEnable = VK_TRUE;
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-    pipe.SetDepthStencil(&ds_ci);
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.layout = pl.handle();
-    create_info.renderPass = VK_NULL_HANDLE;
-
-    VkFormat color_formats = {VK_FORMAT_R8G8B8A8_UNORM};
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-    pipeline_rendering_info.depthAttachmentFormat = VK_FORMAT_D16_UNORM;  // D16_UNORM has guaranteed support
-    pipeline_rendering_info.pNext = create_info.pNext;
-    create_info.pNext = &pipeline_rendering_info;
-
-    const VkRenderPass render_pass = VK_NULL_HANDLE;
-    pipe.CreateVKPipeline(pl.handle(), render_pass, &create_info);
-}
-
-TEST_F(PositiveDynamicRendering, UseStencilAttachmentWithIntegerFormatAndDepthStencilResolve) {
-    TEST_DESCRIPTION("Use stencil attachment with integer format and depth stencil resolve extension");
-    AddRequiredExtensions(VK_KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME);
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-    ASSERT_NO_FATAL_FAILURE(InitRenderTarget());
-
-    if (!ImageFormatIsSupported(gpu(), VK_FORMAT_S8_UINT, VK_IMAGE_TILING_OPTIMAL)) {
-        GTEST_SKIP() << "VK_FORMAT_S8_UINT format not supported";
-    }
-
-    auto depth_stencil_resolve_properties = LvlInitStruct<VkPhysicalDeviceDepthStencilResolveProperties>();
-    GetPhysicalDeviceProperties2(depth_stencil_resolve_properties);
-    if ((depth_stencil_resolve_properties.supportedStencilResolveModes & VK_RESOLVE_MODE_AVERAGE_BIT) == 0) {
-        GTEST_SKIP() << "VK_RESOLVE_MODE_AVERAGE_BIT not supported for VK_FORMAT_S8_UINT";
-    }
-
-    VkImageCreateInfo image_create_info = LvlInitStruct<VkImageCreateInfo>();
-    image_create_info.imageType = VK_IMAGE_TYPE_2D;
-    image_create_info.format = VK_FORMAT_S8_UINT;
-    image_create_info.extent = {32, 32, 1};
-    image_create_info.mipLevels = 1;
-    image_create_info.arrayLayers = 1;
-    image_create_info.samples = VK_SAMPLE_COUNT_4_BIT;
-    image_create_info.tiling = VK_IMAGE_TILING_OPTIMAL;
-    image_create_info.usage = VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-    VkImageObj image(m_device);
-    image.Init(image_create_info);
-    VkImageView image_view = image.targetView(VK_FORMAT_S8_UINT, VK_IMAGE_ASPECT_STENCIL_BIT);
-
-    image_create_info.samples = VK_SAMPLE_COUNT_1_BIT;
-    VkImageObj resolve_image(m_device);
-    resolve_image.Init(image_create_info);
-    VkImageView resolve_image_view = resolve_image.targetView(VK_FORMAT_S8_UINT, VK_IMAGE_ASPECT_STENCIL_BIT);
-
-    VkRenderingAttachmentInfoKHR stencil_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    stencil_attachment.imageView = image_view;
-    stencil_attachment.imageLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-    stencil_attachment.resolveMode = VK_RESOLVE_MODE_AVERAGE_BIT;
-    stencil_attachment.resolveImageLayout = VK_IMAGE_LAYOUT_GENERAL;
-    stencil_attachment.resolveImageView = resolve_image_view;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.pStencilAttachment = &stencil_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, FragmentDensityMapSubsampledBit) {
-    TEST_DESCRIPTION("Test creating an image with subsampled bit.");
-    AddRequiredExtensions(VK_EXT_FRAGMENT_DENSITY_MAP_EXTENSION_NAME);
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkImageCreateInfo image_ci = LvlInitStruct<VkImageCreateInfo>();
-    image_ci.flags = VK_IMAGE_CREATE_SUBSAMPLED_BIT_EXT;
-    image_ci.imageType = VK_IMAGE_TYPE_2D;
-    image_ci.format = VK_FORMAT_R8G8B8A8_UINT;
-    image_ci.extent = {32, 32, 1};
-    image_ci.mipLevels = 1;
-    image_ci.arrayLayers = 1;
-    image_ci.samples = VK_SAMPLE_COUNT_1_BIT;
-    image_ci.tiling = VK_IMAGE_TILING_OPTIMAL;
-    image_ci.usage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-
-    VkImageObj image(m_device);
-    image.Init(image_ci);
-    VkImageView image_view = image.targetView(VK_FORMAT_R8G8B8A8_UINT, VK_IMAGE_ASPECT_COLOR_BIT);
-
-    image_ci.flags = 0;
-    image_ci.usage = VK_IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT;
-
-    VkImageObj fdm_image(m_device);
-    fdm_image.Init(image_ci);
-    VkImageView fdm_image_view = fdm_image.targetView(VK_FORMAT_R8G8B8A8_UINT);
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_GENERAL;
-    color_attachment.imageView = image_view;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-
-    auto fragment_density_map = LvlInitStruct<VkRenderingFragmentDensityMapAttachmentInfoEXT>();
-    fragment_density_map.imageView = fdm_image_view;
-    fragment_density_map.imageLayout = VK_IMAGE_LAYOUT_GENERAL;
-    begin_rendering_info.pNext = &fragment_density_map;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-}
-
-TEST_F(PositiveDynamicRendering, SuspendResumeDraw) {
-    TEST_DESCRIPTION("Resume and suspend at vkCmdBeginRendering time");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_formats = VK_FORMAT_UNDEFINED;
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-
-    pipe.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    VkRect2D scissor = {{0, 0}, {16, 16}};
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.flags = VK_RENDERING_SUSPENDING_BIT_KHR;
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    VkCommandBufferObj cb1(m_device, m_commandPool);
-    VkCommandBufferObj cb2(m_device, m_commandPool);
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-
-    begin_rendering_info.flags = VK_RENDERING_RESUMING_BIT_KHR | VK_RENDERING_SUSPENDING_BIT_KHR;
-    cb1.begin();
-    cb1.BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(cb1.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(cb1.handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(cb1.handle(), 0, 1, &scissor);
-    vk::CmdDraw(cb1.handle(), 3, 1, 0, 0);
-    cb1.EndRendering();
-    cb1.end();
-
-    begin_rendering_info.flags = VK_RENDERING_RESUMING_BIT_KHR;
-    cb2.begin();
-    cb2.BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(cb2.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(cb2.handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(cb2.handle(), 0, 1, &scissor);
-    vk::CmdDraw(cb2.handle(), 3, 1, 0, 0);
-    cb2.EndRendering();
-    cb2.end();
-
-    std::array<VkCommandBuffer, 3> cbs = {m_commandBuffer->handle(), cb1.handle(), cb2.handle()};
-    auto submit = LvlInitStruct<VkSubmitInfo>();
-    submit.commandBufferCount = static_cast<uint32_t>(cbs.size());
-    submit.pCommandBuffers = cbs.data();
-    vk::QueueSubmit(m_device->m_queue, 1, &submit, VK_NULL_HANDLE);
-
-    vk::QueueWaitIdle(m_device->m_queue);
-}
-
-TEST_F(PositiveDynamicRendering, CreateGraphicsPipeline) {
-    TEST_DESCRIPTION("Test for a creating a pipeline with VK_KHR_dynamic_rendering enabled");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    char const* fsSource = R"glsl(
-        #version 450
-        layout(input_attachment_index=0, set=0, binding=0) uniform subpassInput x;
-        layout(location=0) out vec4 color;
-        void main() {
-           color = subpassLoad(x);
-        }
-    )glsl";
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_format = VK_FORMAT_R8G8B8A8_UNORM;
-    auto rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    rendering_info.colorAttachmentCount = 1;
-    rendering_info.pColorAttachmentFormats = &color_format;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &rendering_info;
-
-    VkAttachmentReference attachment = {};
-    attachment.layout = VK_IMAGE_LAYOUT_GENERAL;
-    attachment.attachment = 0;
-
-    VkAttachmentDescription attach_desc = {};
-    attach_desc.format = VK_FORMAT_B8G8R8A8_UNORM;
-    attach_desc.samples = VK_SAMPLE_COUNT_1_BIT;
-    attach_desc.initialLayout = VK_IMAGE_LAYOUT_PREINITIALIZED;
-    attach_desc.finalLayout = VK_IMAGE_LAYOUT_GENERAL;
-
-    VkSubpassDescription subpass = {};
-    subpass.inputAttachmentCount = 1;
-    subpass.pInputAttachments = &attachment;
-    VkRenderPassCreateInfo render_pass_ci = LvlInitStruct<VkRenderPassCreateInfo>();
-    render_pass_ci.subpassCount = 1;
-    render_pass_ci.pSubpasses = &subpass;
-    render_pass_ci.attachmentCount = 1;
-    render_pass_ci.pAttachments = &attach_desc;
-
-    vk_testing::RenderPass render_pass;
-    render_pass.init(*m_device, render_pass_ci);
-
-    pipe.CreateVKPipeline(pl.handle(), render_pass.handle(), &create_info);
-}
-
-TEST_F(PositiveDynamicRendering, CreateGraphicsPipelineNoInfo) {
-    TEST_DESCRIPTION("Test for a creating a pipeline with VK_KHR_dynamic_rendering enabled but no rendering info struct.");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    char const* fsSource = R"glsl(
-        #version 450
-        layout(input_attachment_index=0, set=0, binding=0) uniform subpassInput x;
-        layout(location=0) out vec4 color;
-        void main() {
-           color = subpassLoad(x);
-        }
-    )glsl";
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkAttachmentReference attachment = {};
-    attachment.layout = VK_IMAGE_LAYOUT_GENERAL;
-    attachment.attachment = 0;
-
-    VkAttachmentDescription attach_desc = {};
-    attach_desc.format = VK_FORMAT_B8G8R8A8_UNORM;
-    attach_desc.samples = VK_SAMPLE_COUNT_1_BIT;
-    attach_desc.initialLayout = VK_IMAGE_LAYOUT_PREINITIALIZED;
-    attach_desc.finalLayout = VK_IMAGE_LAYOUT_GENERAL;
-
-    VkSubpassDescription subpass = {};
-    subpass.inputAttachmentCount = 1;
-    subpass.pInputAttachments = &attachment;
-    VkRenderPassCreateInfo render_pass_ci = LvlInitStruct<VkRenderPassCreateInfo>();
-    render_pass_ci.subpassCount = 1;
-    render_pass_ci.pSubpasses = &subpass;
-    render_pass_ci.attachmentCount = 1;
-    render_pass_ci.pAttachments = &attach_desc;
-
-    vk_testing::RenderPass render_pass;
-    render_pass.init(*m_device, render_pass_ci);
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    // if there isn't a VkPipelineRenderingCreateInfoKHR, the driver is supposed to use safe default values
-    pipe.CreateVKPipeline(pl.handle(), render_pass.handle(), &create_info);
-}
-
-TEST_F(PositiveDynamicRendering, CommandDrawWithShaderTileImageRead) {
-    TEST_DESCRIPTION("vkCmdDraw* with shader tile image read extension using dynamic Rendering Tests.");
-
-    SetTargetApiVersion(VK_API_VERSION_1_3);
-
-    AddRequiredExtensions(VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME);
-    AddRequiredExtensions(VK_EXT_SHADER_TILE_IMAGE_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_3) {
-        GTEST_SKIP() << "At least Vulkan version 1.3 is required";
-    }
-
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto dynamic_rendering_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingFeaturesKHR>();
-    auto shader_tile_image_features = LvlInitStruct<VkPhysicalDeviceShaderTileImageFeaturesEXT>();
-    dynamic_rendering_features.pNext = &shader_tile_image_features;
-    auto features2 = GetPhysicalDeviceFeatures2(dynamic_rendering_features);
-    if (!dynamic_rendering_features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering";
-    }
-
-    // shader tile image read features not supported skip the test.
-    if (!shader_tile_image_features.shaderTileImageDepthReadAccess &&
-        !shader_tile_image_features.shaderTileImageStencilReadAccess) {
-        GTEST_SKIP() << "Test requires (unsupported) shader tile image extension.";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    auto fs =
-        VkShaderObj::CreateFromASM(*this, VK_SHADER_STAGE_FRAGMENT_BIT, bindShaderTileImageDepthStencilReadSpv, "main", nullptr);
-
-    const VkDynamicState dyn_states[] = {VK_DYNAMIC_STATE_DEPTH_WRITE_ENABLE, VK_DYNAMIC_STATE_STENCIL_WRITE_MASK};
-    auto dyn_state_ci = LvlInitStruct<VkPipelineDynamicStateCreateInfo>();
-    dyn_state_ci.dynamicStateCount = 2;
-    dyn_state_ci.pDynamicStates = dyn_states;
-
-    auto ds_state = LvlInitStruct<VkPipelineDepthStencilStateCreateInfo>();
-    ds_state.depthWriteEnable = VK_TRUE;
-
-    VkFormat depth_format = VK_FORMAT_D32_SFLOAT_S8_UINT;
-    VkFormat color_format = VK_FORMAT_B8G8R8A8_UNORM;
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_format;
-    pipeline_rendering_info.depthAttachmentFormat = depth_format;
-    pipeline_rendering_info.stencilAttachmentFormat = depth_format;
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-
-    auto ms_ci = LvlInitStruct<VkPipelineMultisampleStateCreateInfo>();
-    ms_ci.sampleShadingEnable = VK_TRUE;
-    ms_ci.minSampleShading = 1.0;
-    ms_ci.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-
-    CreatePipelineHelper pipe(*this);
-    pipe.InitInfo();
-    pipe.shader_stages_ = {vs.GetStageCreateInfo(), fs->GetStageCreateInfo()};
-    pipe.gp_ci_.pNext = &pipeline_rendering_info;
-    pipe.gp_ci_.renderPass = VK_NULL_HANDLE;
-    pipe.gp_ci_.pMultisampleState = &ms_ci;
-    pipe.gp_ci_.pDepthStencilState = &ds_state;
-    pipe.dyn_state_ci_ = dyn_state_ci;
-    pipe.pipeline_layout_ = VkPipelineLayoutObj(m_device, {&dsl});
-    pipe.InitState();
-    pipe.CreateGraphicsPipeline();
-
-    VkImageObj depth_image(m_device);
-    depth_image.Init(32, 32, 1, depth_format, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, VK_IMAGE_TILING_OPTIMAL, 0);
-    ASSERT_TRUE(depth_image.initialized());
-
-    VkImageViewCreateInfo depth_view_ci = {VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
-                                           nullptr,
-                                           0,
-                                           depth_image.handle(),
-                                           VK_IMAGE_VIEW_TYPE_2D,
-                                           depth_format,
-                                           {VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY,
-                                            VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY},
-                                           {VK_IMAGE_ASPECT_DEPTH_BIT | VK_IMAGE_ASPECT_STENCIL_BIT, 0, 1, 0, 1}};
-
-    vk_testing::ImageView depth_image_view(*m_device, depth_view_ci);
-
-    VkImageObj color_image(m_device);
-    color_image.Init(32, 32, 1, color_format, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_TILING_OPTIMAL, 0);
-    ASSERT_TRUE(color_image.initialized());
-
-    VkImageViewCreateInfo color_view_ci = {VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
-                                           nullptr,
-                                           0,
-                                           color_image.handle(),
-                                           VK_IMAGE_VIEW_TYPE_2D,
-                                           color_format,
-                                           {VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY,
-                                            VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY},
-                                           {VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1}};
-
-    vk_testing::ImageView color_image_view(*m_device, color_view_ci);
-
-    VkRenderingAttachmentInfoKHR depth_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    depth_attachment.imageLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-    depth_attachment.imageView = depth_image_view.handle();
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    color_attachment.imageView = color_image_view.handle();
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.pDepthAttachment = &depth_attachment;
-    begin_rendering_info.pStencilAttachment = &depth_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.pipeline_);
-    vk::CmdSetDepthWriteEnable(m_commandBuffer->handle(), false);
-    vk::CmdSetStencilWriteMask(m_commandBuffer->handle(), VK_STENCIL_FACE_FRONT_BIT, 0);
-    vk::CmdSetStencilWriteMask(m_commandBuffer->handle(), VK_STENCIL_FACE_BACK_BIT, 0);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_errorMonitor->VerifyFound();
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, DualSourceBlending) {
-    TEST_DESCRIPTION("Test drawing with dynamic rendering and dual source blending.");
-
-    SetTargetApiVersion(VK_API_VERSION_1_1);
-    AddRequiredExtensions(VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_1) {
-        GTEST_SKIP() << "At least Vulkan version 1.1 is required";
-    }
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto dynamic_rendering_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingFeaturesKHR>();
-    auto features2 = GetPhysicalDeviceFeatures2(dynamic_rendering_features);
-    if (features2.features.dualSrcBlend == VK_FALSE) {
-        GTEST_SKIP() << "Test requires (unsupported) dualSrcBlend";
-    }
-    if (dynamic_rendering_features.dynamicRendering == VK_FALSE) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    char const* fsSource = R"glsl(
-        #version 450
-        layout(location = 0) out vec4 c0;
-        void main(){
-           c0 = vec4(1.0f);
-        }
-    )glsl";
-
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineColorBlendAttachmentState cb_attachments = {};
-    cb_attachments.blendEnable = VK_TRUE;
-    cb_attachments.srcColorBlendFactor = VK_BLEND_FACTOR_SRC1_COLOR;  // bad!
-    cb_attachments.dstColorBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_COLOR;
-    cb_attachments.colorBlendOp = VK_BLEND_OP_ADD;
-    cb_attachments.srcAlphaBlendFactor = VK_BLEND_FACTOR_SRC_ALPHA;
-    cb_attachments.dstAlphaBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-    cb_attachments.alphaBlendOp = VK_BLEND_OP_ADD;
-
-    VkFormat color_formats = VK_FORMAT_UNDEFINED;
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-
-    CreatePipelineHelper pipe(*this);
-    pipe.InitInfo();
-    pipe.gp_ci_.pNext = &pipeline_rendering_info;
-    pipe.cb_attachments_[0] = cb_attachments;
-    pipe.shader_stages_ = {pipe.vs_->GetStageCreateInfo(), fs.GetStageCreateInfo()};
-    pipe.InitState();
-    pipe.CreateGraphicsPipeline();
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.pipeline_);
-
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, ExecuteCommandsWithNullImageView) {
-    TEST_DESCRIPTION(
-        "Test CmdExecuteCommands with an inherited image format of VK_FORMAT_UNDEFINED inside a render pass begun with "
-        "CmdBeginRendering where the same image is specified as null");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    auto color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    color_attachment.imageView = VK_NULL_HANDLE;
-
-    constexpr std::array bad_color_formats = {VK_FORMAT_UNDEFINED};
-
-    auto inheritance_rendering_info = LvlInitStruct<VkCommandBufferInheritanceRenderingInfoKHR>();
-    inheritance_rendering_info.colorAttachmentCount = bad_color_formats.size();
-    inheritance_rendering_info.pColorAttachmentFormats = bad_color_formats.data();
-    inheritance_rendering_info.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-
-    auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.flags = VK_RENDERING_CONTENTS_SECONDARY_COMMAND_BUFFERS_BIT_KHR;
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    VkCommandBufferObj secondary(m_device, m_commandPool, VK_COMMAND_BUFFER_LEVEL_SECONDARY);
-
-    const auto cmdbuff_ii = LvlInitStruct<VkCommandBufferInheritanceInfo>(&inheritance_rendering_info);
-
-    auto cmdbuff_bi = LvlInitStruct<VkCommandBufferBeginInfo>();
-    cmdbuff_bi.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT | VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT;
-    cmdbuff_bi.pInheritanceInfo = &cmdbuff_ii;
-
-    secondary.begin(&cmdbuff_bi);
-    secondary.end();
-
-    m_commandBuffer->begin();
-
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-
-    vk::CmdExecuteCommands(m_commandBuffer->handle(), 1, &secondary.handle());
-
-    m_commandBuffer->EndRendering();
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, SuspendPrimaryResumeInSecondary) {
-    TEST_DESCRIPTION("Suspend in primary and resume in secondary");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_formats = {VK_FORMAT_UNDEFINED};
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-
-    pipe.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    VkRect2D scissor = {{0, 0}, {16, 16}};
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.flags = VK_RENDERING_SUSPENDING_BIT_KHR;
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    // Primary suspends render
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-
-    // Secondary resumes render
-    VkCommandBufferObj secondary(m_device, m_commandPool, VK_COMMAND_BUFFER_LEVEL_SECONDARY);
-    begin_rendering_info.flags = VK_RENDERING_RESUMING_BIT_KHR;
-    secondary.begin();
-    secondary.BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(secondary.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(secondary.handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(secondary.handle(), 0, 1, &scissor);
-    secondary.Draw(3, 1, 0, 0);
-    secondary.EndRendering();
-    secondary.end();
-
-    // Execute secondary
-    vk::CmdExecuteCommands(m_commandBuffer->handle(), 1, &secondary.handle());
-
-    m_commandBuffer->end();
-
-    // Submit
-    auto submit = LvlInitStruct<VkSubmitInfo>();
-    submit.commandBufferCount = 1;
-    submit.pCommandBuffers = &m_commandBuffer->handle();
-    vk::QueueSubmit(m_device->m_queue, 1, &submit, VK_NULL_HANDLE);
-
-    vk::QueueWaitIdle(m_device->m_queue);
-}
-
-TEST_F(PositiveDynamicRendering, SuspendSecondaryResumeInPrimary) {
-    TEST_DESCRIPTION("Suspend in secondary and resume in primary");
-    InitBasicDynamicRendering();
-    if (::testing::Test::IsSkipped()) return;
-
-    char const* fsSource = R"glsl(
-        #version 450
-        layout(location=0) out vec4 color;
-        void main() {
-           color = vec4(1.0f);
-        }
-    )glsl";
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.AddDefaultColorAttachment();
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    VkFormat color_formats = {VK_FORMAT_UNDEFINED};
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &color_formats;
-
-    auto create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-    pipe.InitGraphicsPipelineCreateInfo(&create_info);
-    create_info.pNext = &pipeline_rendering_info;
-
-    pipe.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &create_info);
-
-    VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    VkRect2D scissor = {{0, 0}, {16, 16}};
-
-    VkRenderingAttachmentInfoKHR color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    begin_rendering_info.flags = 0;
-    begin_rendering_info.colorAttachmentCount = 1;
-    begin_rendering_info.pColorAttachments = &color_attachment;
-    begin_rendering_info.layerCount = 1;
-
-    // First primary with secondary that suspends render
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &scissor);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRendering();
-
-    VkCommandBufferObj secondary(m_device, m_commandPool, VK_COMMAND_BUFFER_LEVEL_SECONDARY);
-    begin_rendering_info.flags = VK_RENDERING_SUSPENDING_BIT_KHR;
-    secondary.begin();
-    secondary.BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(secondary.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(secondary.handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(secondary.handle(), 0, 1, &scissor);
-    secondary.Draw(3, 1, 0, 0);
-    secondary.EndRendering();
-    secondary.end();
-
-    // Execute secondary
-    vk::CmdExecuteCommands(m_commandBuffer->handle(), 1, &secondary.handle());
-
-    m_commandBuffer->end();
-
-    // Second Primary resumes render
-    VkCommandBufferObj cb(m_device, m_commandPool);
-    begin_rendering_info.flags = VK_RENDERING_RESUMING_BIT_KHR;
-    cb.begin();
-    cb.BeginRendering(begin_rendering_info);
-    vk::CmdBindPipeline(cb.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-    vk::CmdSetViewport(cb.handle(), 0, 1, &viewport);
-    vk::CmdSetScissor(cb.handle(), 0, 1, &scissor);
-    vk::CmdDraw(cb.handle(), 3, 1, 0, 0);
-    cb.EndRendering();
-    cb.end();
-
-    // Submit
-    std::array<VkCommandBuffer, 2> cbs = {m_commandBuffer->handle(), cb.handle()};
-    auto submit = LvlInitStruct<VkSubmitInfo>();
-    submit.commandBufferCount = static_cast<uint32_t>(cbs.size());
-    submit.pCommandBuffers = cbs.data();
-    vk::QueueSubmit(m_device->m_queue, 1, &submit, VK_NULL_HANDLE);
-
-    vk::QueueWaitIdle(m_device->m_queue);
-}
-
-TEST_F(PositiveDynamicRendering, WithShaderTileImageAndBarrier) {
-    TEST_DESCRIPTION("Test setting memory barrier with shader tile image features are enabled.");
-
-    SetTargetApiVersion(VK_API_VERSION_1_3);
-    AddRequiredExtensions(VK_EXT_SHADER_TILE_IMAGE_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_3) {
-        GTEST_SKIP() << "At least Vulkan version 1.3 is required";
-    }
-
-    auto vk13features = LvlInitStruct<VkPhysicalDeviceVulkan13Features>();
-    auto shader_tile_image_features = LvlInitStruct<VkPhysicalDeviceShaderTileImageFeaturesEXT>();
-    vk13features.pNext = &shader_tile_image_features;
-
-    auto features2 = GetPhysicalDeviceFeatures2(vk13features);
-    if (!vk13features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering";
-    }
-    if (!vk13features.synchronization2) {
-        GTEST_SKIP() << "Test requires (unsupported) synchronization2";
-    }
-
-    if (!shader_tile_image_features.shaderTileImageColorReadAccess && !shader_tile_image_features.shaderTileImageDepthReadAccess &&
-        !shader_tile_image_features.shaderTileImageStencilReadAccess) {
-        GTEST_SKIP() << "Test requires (unsupported) shader tile image extension.";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-    ASSERT_NO_FATAL_FAILURE(InitRenderTarget());
-
-    m_commandBuffer->begin();
-
-    VkRenderingInfoKHR begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-    VkClearRect clear_rect = {{{0, 0}, {m_width, m_height}}, 0, 1};
-    begin_rendering_info.renderArea = clear_rect.rect;
-    begin_rendering_info.layerCount = 1;
-
-    auto memory_barrier_2 = LvlInitStruct<VkMemoryBarrier2KHR>();
-    memory_barrier_2.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-    memory_barrier_2.srcAccessMask = VK_ACCESS_2_COLOR_ATTACHMENT_WRITE_BIT;
-    memory_barrier_2.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-    memory_barrier_2.dstAccessMask = VK_ACCESS_2_COLOR_ATTACHMENT_READ_BIT;
-
-    auto dependency_info = LvlInitStruct<VkDependencyInfoKHR>();
-    dependency_info.dependencyFlags = VK_DEPENDENCY_BY_REGION_BIT;
-    dependency_info.memoryBarrierCount = 1;
-    dependency_info.pMemoryBarriers = &memory_barrier_2;
-    dependency_info.bufferMemoryBarrierCount = 0;
-    dependency_info.pBufferMemoryBarriers = VK_NULL_HANDLE;
-    dependency_info.imageMemoryBarrierCount = 0;
-    dependency_info.pImageMemoryBarriers = VK_NULL_HANDLE;
-
-    vk::CmdBeginRendering(m_commandBuffer->handle(), &begin_rendering_info);
-    vk::CmdPipelineBarrier2(m_commandBuffer->handle(), &dependency_info);
-    vk::CmdEndRendering(m_commandBuffer->handle());
-
-    auto memory_barrier = LvlInitStruct<VkMemoryBarrier>();
-    memory_barrier.srcAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-    memory_barrier.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_READ_BIT;
-    vk::CmdBeginRendering(m_commandBuffer->handle(), &begin_rendering_info);
-    vk::CmdPipelineBarrier(m_commandBuffer->handle(), VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT,
-                           VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT, VK_DEPENDENCY_BY_REGION_BIT, 1, &memory_barrier, 0,
-                           nullptr, 0, nullptr);
-    vk::CmdEndRendering(m_commandBuffer->handle());
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, MatchingAttachmentFormats) {
-    TEST_DESCRIPTION(
-        "Draw with Dynamic Rendering with attachment specified as VK_NULL_HANDLE in VkRenderingInfoKHR, and with corresponding "
-        "format in VkPipelineRenderingCreateInfoKHR set to VK_FORMAT_UNDEFINED");
-
-    SetTargetApiVersion(VK_API_VERSION_1_1);
-
-    AddRequiredExtensions(VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_1) {
-        GTEST_SKIP() << "At least Vulkan version 1.1 is required";
-    }
-
-    auto dynamic_rendering_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingFeaturesKHR>();
-    auto features2 = GetPhysicalDeviceFeatures2(dynamic_rendering_features);
-    if (!dynamic_rendering_features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    const VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    const VkRect2D scissor = {{0, 0}, {16, 16}};
-    m_viewports.push_back(viewport);
-    m_scissors.push_back(scissor);
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-
-    VkPipelineObj pipeline_color(m_device);
-    pipeline_color.AddShader(&vs);
-    pipeline_color.AddShader(&fs);
-    pipeline_color.AddDefaultColorAttachment();
-    pipeline_color.SetViewport(m_viewports);
-    pipeline_color.SetScissor(m_scissors);
-
-    VkFormat color_formats[] = {VK_FORMAT_UNDEFINED};
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = color_formats;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_color.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_color.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-    auto ds_state = LvlInitStruct<VkPipelineDepthStencilStateCreateInfo>();
-
-    VkPipelineObj pipeline_depth(m_device);
-    pipeline_depth.AddShader(&vs);
-    pipeline_depth.AddShader(&fs);
-    pipeline_depth.AddDefaultColorAttachment();
-    pipeline_depth.SetViewport(m_viewports);
-    pipeline_depth.SetScissor(m_scissors);
-    pipeline_depth.SetDepthStencil(&ds_state);
-
-    pipeline_rendering_info.colorAttachmentCount = 0;
-    pipeline_rendering_info.pColorAttachmentFormats = nullptr;
-    pipeline_rendering_info.depthAttachmentFormat = VK_FORMAT_UNDEFINED;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_depth.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_depth.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-
-    VkPipelineObj pipeline_stencil(m_device);
-    pipeline_stencil.AddShader(&vs);
-    pipeline_stencil.AddShader(&fs);
-    pipeline_stencil.AddDefaultColorAttachment();
-    pipeline_stencil.SetViewport(m_viewports);
-    pipeline_stencil.SetScissor(m_scissors);
-    pipeline_stencil.SetDepthStencil(&ds_state);
-
-    pipeline_rendering_info.colorAttachmentCount = 0;
-    pipeline_rendering_info.depthAttachmentFormat = VK_FORMAT_UNDEFINED;
-    pipeline_rendering_info.stencilAttachmentFormat = VK_FORMAT_UNDEFINED;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_stencil.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_stencil.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-
-    VkImageObj colorImage(m_device);
-    colorImage.Init(32, 32, 1, VK_FORMAT_R8G8B8A8_UINT, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT);
-
-    VkImageObj depthStencilImage(m_device);
-    const VkFormat depthStencilFormat = FindSupportedDepthStencilFormat(gpu());
-    depthStencilImage.Init(32, 32, 1, depthStencilFormat, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT);
-
-    auto color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    color_attachment.imageView = VK_NULL_HANDLE;
-
-    auto depth_stencil_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    depth_stencil_attachment.imageLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-    depth_stencil_attachment.imageView = VK_NULL_HANDLE;
-
-    m_commandBuffer->begin();
-
-    {
-        // Mismatching color formats
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.colorAttachmentCount = 1;
-        begin_rendering_info.pColorAttachments = &color_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_color.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-    }
-
-    {
-        // Mismatching depth format
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.pDepthAttachment = &depth_stencil_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_depth.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-    }
-
-    {
-        // Mismatching stencil format
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.pStencilAttachment = &depth_stencil_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_stencil.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-    }
-
-    m_commandBuffer->end();
-}
-
-TEST_F(PositiveDynamicRendering, MatchingAttachmentFormats2) {
-    TEST_DESCRIPTION(
-        "Draw with Dynamic Rendering with dynamicRenderingUnusedAttachments enabled and matching formats");
-
-    SetTargetApiVersion(VK_API_VERSION_1_2);
-    AddRequiredExtensions(VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME);
-    AddRequiredExtensions(VK_EXT_DYNAMIC_RENDERING_UNUSED_ATTACHMENTS_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (DeviceValidationVersion() < VK_API_VERSION_1_2) {
-        GTEST_SKIP() << "At least Vulkan version 1.2 is required";
-    }
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto dynamic_rendering_unused_attachments_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingUnusedAttachmentsFeaturesEXT>();
-    auto dynamic_rendering_features = LvlInitStruct<VkPhysicalDeviceDynamicRenderingFeaturesKHR>(&dynamic_rendering_unused_attachments_features);
-    GetPhysicalDeviceFeatures2(dynamic_rendering_features);
-    if (!dynamic_rendering_features.dynamicRendering) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRendering , skipping.";
-    }
-    if (!dynamic_rendering_unused_attachments_features.dynamicRenderingUnusedAttachments) {
-        GTEST_SKIP() << "Test requires (unsupported) dynamicRenderingUnusedAttachments , skipping.";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &dynamic_rendering_features, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT));
-
-    const VkViewport viewport = {0, 0, 16, 16, 0, 1};
-    const VkRect2D scissor = {{0, 0}, {16, 16}};
-    m_viewports.push_back(viewport);
-    m_scissors.push_back(scissor);
-
-    VkShaderObj vs(this, bindStateVertShaderText, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, bindStateFragColorOutputText, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkDescriptorSetLayoutBinding dslb = {0, VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj dsl(m_device, {dslb});
-    const VkPipelineLayoutObj pl(m_device, {&dsl});
-
-    auto pipeline_rendering_info = LvlInitStruct<VkPipelineRenderingCreateInfoKHR>();
-
-    VkPipelineObj pipeline_color(m_device);
-    pipeline_color.AddShader(&vs);
-    pipeline_color.AddShader(&fs);
-    pipeline_color.AddDefaultColorAttachment();
-    pipeline_color.SetViewport(m_viewports);
-    pipeline_color.SetScissor(m_scissors);
-
-    VkFormat color_formats[] = {VK_FORMAT_R8G8B8A8_UNORM};
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = color_formats;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_color.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_color.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-    auto ds_state = LvlInitStruct<VkPipelineDepthStencilStateCreateInfo>();
-
-    VkFormat depthStencilFormat = FindSupportedDepthStencilFormat(gpu());
-
-    VkPipelineObj pipeline_depth(m_device);
-    pipeline_depth.AddShader(&vs);
-    pipeline_depth.AddShader(&fs);
-    pipeline_depth.AddDefaultColorAttachment();
-    pipeline_depth.SetViewport(m_viewports);
-    pipeline_depth.SetScissor(m_scissors);
-    pipeline_depth.SetDepthStencil(&ds_state);
-
-    pipeline_rendering_info.colorAttachmentCount = 0;
-    pipeline_rendering_info.pColorAttachmentFormats = nullptr;
-    pipeline_rendering_info.depthAttachmentFormat = depthStencilFormat;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_depth.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_depth.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-
-    VkPipelineObj pipeline_stencil(m_device);
-    pipeline_stencil.AddShader(&vs);
-    pipeline_stencil.AddShader(&fs);
-    pipeline_stencil.AddDefaultColorAttachment();
-    pipeline_stencil.SetViewport(m_viewports);
-    pipeline_stencil.SetScissor(m_scissors);
-    pipeline_stencil.SetDepthStencil(&ds_state);
-
-    pipeline_rendering_info.colorAttachmentCount = 0;
-    pipeline_rendering_info.depthAttachmentFormat = VK_FORMAT_UNDEFINED;
-    pipeline_rendering_info.stencilAttachmentFormat = depthStencilFormat;
-
-    {
-        auto pipeline_create_info = LvlInitStruct<VkGraphicsPipelineCreateInfo>();
-        pipeline_stencil.InitGraphicsPipelineCreateInfo(&pipeline_create_info);
-        pipeline_create_info.pNext = &pipeline_rendering_info;
-
-        pipeline_stencil.CreateVKPipeline(pl.handle(), VK_NULL_HANDLE, &pipeline_create_info);
-    }
-
-    VkImageObj colorImage(m_device);
-    colorImage.Init(32, 32, 1, VK_FORMAT_R8G8B8A8_UNORM, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT);
-
-    VkImageObj depthStencilImage(m_device);
-    depthStencilImage.Init(32, 32, 1, depthStencilFormat, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT);
-
-    auto color_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    color_attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    color_attachment.imageView = VK_NULL_HANDLE;
-
-    auto depth_stencil_attachment = LvlInitStruct<VkRenderingAttachmentInfoKHR>();
-    depth_stencil_attachment.imageLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-    depth_stencil_attachment.imageView = VK_NULL_HANDLE;
-
-    m_commandBuffer->begin();
-
-    {
-        // Null color image view
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.colorAttachmentCount = 1;
-        begin_rendering_info.pColorAttachments = &color_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_color.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-
-        // Matching color formats
-        color_attachment.imageView = colorImage.targetView(VK_FORMAT_R8G8B8A8_UNORM);
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_color.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-        color_attachment.imageView = VK_NULL_HANDLE;
-    }
-
-    {
-        // Null depth image view
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.pDepthAttachment = &depth_stencil_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_depth.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-
-        // Matching depth format
-        depth_stencil_attachment.imageView = depthStencilImage.targetView(depthStencilFormat, VK_IMAGE_ASPECT_DEPTH_BIT);
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_depth.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-        depth_stencil_attachment.imageView = VK_NULL_HANDLE;
-    }
-
-    {
-        // Null stencil image view
-        auto begin_rendering_info = LvlInitStruct<VkRenderingInfoKHR>();
-        begin_rendering_info.layerCount = 1;
-        begin_rendering_info.pStencilAttachment = &depth_stencil_attachment;
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_stencil.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-
-        // Matching stencil format
-        depth_stencil_attachment.imageView = depthStencilImage.targetView(depthStencilFormat, VK_IMAGE_ASPECT_STENCIL_BIT);
-        m_commandBuffer->BeginRendering(begin_rendering_info);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_stencil.handle());
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRendering();
-        depth_stencil_attachment.imageView = VK_NULL_HANDLE;
-    }
-
-    m_commandBuffer->end();
-}
